+++ src/cheat/features/FeatureManager.h
@@ -30,29 +30,71 @@
 #define FEATUREMANAGER_H
 #pragma once
 
+enum EFeatureType
+{
+	FEATURE_Unset,
+
+	FEATURE_Bool,
+	FEATURE_Integer,
+	FEATURE_Float,
+};
+
 class BaseFeature
 {
 public:
 	BaseFeature(VarBoolean* var, const std::string& name)
 	{
-		m_variable = var;
+		m_type = FEATURE_Bool;
+		m_base_variable = m_variable_bool = var;
+		m_name = name;
+	}
+
+	BaseFeature(VarInteger* var, int off_value, const std::string& name)
+	{
+		m_type = FEATURE_Integer;
+		m_base_variable = m_variable_integer = var;
+		m_off_value_int = off_value;
+		m_name = name;
+	}
+
+	BaseFeature(VarFloat* var, float off_value, const std::string& name)
+	{
+		m_type = FEATURE_Float;
+		m_base_variable = m_variable_float = var;
+		m_off_value_float = off_value;
 		m_name = name;
 	}
 
 	const std::string& get_name() const { return m_name; }
-	VarBoolean* get_variable() const { assert(m_variable); return m_variable; }
+	BaseVariable* get_variable() const { assert(m_base_variable); return m_base_variable; }
 
 	bool is_enabled() const
 	{
-		return m_variable->get_value();
+		switch (m_type)
+		{
+			case FEATURE_Bool: return m_variable_bool->get_value() != false;
+			case FEATURE_Integer: return m_variable_integer->get_value() != m_off_value_int;
+			case FEATURE_Float: return m_variable_float->get_value() != m_off_value_float;
+		}
+
+		assert(0 && "Unknown feature type.");
+		return false;
 	};
 
 private:
 	std::string m_name = "unset";
 
-	VarBoolean* m_variable = nullptr;
+	VarBoolean* m_variable_bool = nullptr;
+	VarInteger* m_variable_integer = nullptr;
+	int m_off_value_int = 0;
+	VarFloat* m_variable_float = nullptr;
+	float m_off_value_float = 0;
+
+	BaseVariable* m_base_variable = nullptr;
 
 	std::function<bool(VarBoolean*)> m_custom_is_enabled;
+
+	EFeatureType m_type = FEATURE_Unset;
 };
 
 class CFeatureManager
@@ -65,8 +107,13 @@ public:
 
 	void add_new_feature(const BaseFeature& feature);
 
-	void for_each_feature(void(*callback)(BaseFeature* feature))
+	void for_each_feature(const std::function<void(BaseFeature* f)>& callback)
 	{
+		if (!m_initialized)
+		{
+			return;
+		}
+
 		for (auto& f : m_features)
 		{
 			callback(&f);
@@ -78,6 +125,8 @@ private:
 
 private:
 	std::vector<BaseFeature> m_features;
+
+	bool m_initialized = false;
 };
 
 #endif // FEATUREMANAGER_H